requirements.txt
Flask==2.0.3
Flask_Cors==3.0.10
Flask_SQLAlchemy==2.5.1
SQLAlchemy==1.4.31
python-dotenv==0.19.2
psycopg2-binary
PyJWT
requests